lambda_functions/file_upload_processor/requirements.txt
boto3
requests
weasyprint
awslambdaric
markdown-it-py
pandas
openpyxl